feat!: Run container directly as unprivileged user
This avoids dropping privileges from root to a regular user as part of
the container initialization.

To achieve this, `/etc/webhook.yml` must be moved to a location writeable by
the unprivileged user. This is now `/etc/voxpupuli/webhook/webhook.yml`,
one of the default locations of webhook-go.

As the container starts directly as Puppet, `webhook-go` is now PID 1 as
well and responds to signals.